+++ src/neptune/api/searching_entries.py
@@ -23,8 +23,6 @@ from typing import (
     Iterable,
     List,
     Optional,
-    Type,
-    Union,
 )
 
 from bravado.client import construct_request  # type: ignore
@@ -45,9 +43,9 @@ from neptune.internal.backends.nql import (
     NQLQueryAggregate,
     NQLQueryAttribute,
 )
-from neptune.internal.backends.utils import which_progress_bar
+from neptune.internal.backends.utils import construct_progress_bar
 from neptune.internal.init.parameters import MAX_SERVER_OFFSET
-from neptune.typing import ProgressBarCallback
+from neptune.typing import ProgressBarType
 
 if TYPE_CHECKING:
     from neptune.internal.backends.swagger_client_wrapper import SwaggerClientWrapper
@@ -150,7 +148,7 @@ def iter_over_pages(
     max_offset: int = MAX_SERVER_OFFSET,
     sort_by_column_type: Optional[str] = None,
     ascending: bool = False,
-    progress_bar: Optional[Union[bool, Type[ProgressBarCallback]]] = None,
+    progress_bar: Optional[ProgressBarType] = None,
     **kwargs: Any,
 ) -> Generator[Any, None, None]:
     searching_after = None
@@ -160,7 +158,7 @@ def iter_over_pages(
 
     progress_bar = progress_bar if step_size >= total else None
 
-    with _construct_progress_bar(progress_bar) as bar:
+    with construct_progress_bar(progress_bar, "Fetching table...") as bar:
         # beginning of the first page
         bar.update(
             by=0,
@@ -205,11 +203,6 @@ def iter_over_pages(
                 yield from page
 
                 last_page = page
-
-
-def _construct_progress_bar(progress_bar: Optional[Union[bool, Type[ProgressBarCallback]]]) -> ProgressBarCallback:
-    progress_bar_type = which_progress_bar(progress_bar)
-    return progress_bar_type(description="Fetching table...")
 
 
 def _entries_from_page(single_page: Dict[str, Any]) -> List[LeaderboardEntry]: